% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPRESSO.cv.R
\name{cv.eval}
\alias{cv.eval}
\title{Title}
\usage{
cv.eval(res.lasso, u.vec, v.vec, cov.mat, n)
}
\arguments{
\item{res.lasso}{}

\item{u.vec}{}

\item{v.vec}{}

\item{cov.mat}{}

\item{n}{}
}
\description{
Title
}
