% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPRESSO.cv.R
\name{cv.lassosum}
\alias{cv.lassosum}
\title{Title}
\usage{
cv.lassosum(r.t, r.v, R.t, R.v, n.t, n.v, lambda_list_1, group, maxIter, phi)
}
\arguments{
\item{r.t}{}

\item{r.v}{}

\item{R.t}{}

\item{R.v}{}

\item{n.t}{}

\item{n.v}{}

\item{lambda_list_1}{}

\item{group}{}

\item{maxIter}{}

\item{phi}{}
}
\description{
Title
}
