% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPRESSO.cv.R
\name{EXPRESSO.cv}
\alias{EXPRESSO.cv}
\title{Title}
\usage{
EXPRESSO.cv(b.vec, cor.analysis, ss, n, group, maxIter, fold)
}
\arguments{
\item{b.vec}{}

\item{cor.analysis}{}

\item{ss}{}

\item{n}{}

\item{group}{}

\item{maxIter}{}

\item{fold}{}
}
\description{
Title
}
