% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPRESSO.final.R
\name{EXPRESSO}
\alias{EXPRESSO}
\title{Title}
\usage{
EXPRESSO(
  sumstatFile,
  annoFile,
  windowFile,
  refFile,
  out_path,
  minMaf = 0.05,
  refIndv = NULL,
  maxIter = 1000,
  gene.vec = NULL,
  append = TRUE
)
}
\arguments{
\item{sumstatFile}{}

\item{annoFile}{}

\item{windowFile}{}

\item{refFile}{}

\item{out_path}{}

\item{minMaf}{}

\item{refIndv}{}

\item{maxIter}{}

\item{gene.vec}{}

\item{append}{}
}
\description{
Title
}
