% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPRESSO.final.R
\name{getVarInWindow}
\alias{getVarInWindow}
\title{get windows;}
\usage{
getVarInWindow(pos.var, gene.window)
}
\arguments{
\item{pos.var}{the variant positions surrouding the transcript;}

\item{gene.window}{the windows defined for the gene;}
}
\description{
get windows;
}
